requirements.txt
Django==1.7.4
django-filter==0.9.2
djangorestframework==3.0.5
Markdown==2.6
Pillow==2.7.0
factory-boy==2.4.1
django-vote==1.2
django-rest-swagger==0.2.8
